dapp-refactor: moved prepareInfoValuesToShow() and TransactionResultTable() to its main shared folder
