+++ system-contract-dapp-playground/src/components/contract-interaction/hts/shared/components/TokenSpecificInfoModal.tsx
@@ -21,7 +21,7 @@
 import Link from 'next/link';
 import { Dispatch, SetStateAction } from 'react';
 import { convertCalmelCaseFunctionName } from '@/utils/common/helpers';
-import { prepareInfoValuesToShow } from '../methods/prepareInfoValuesToShow';
+import { prepareInfoValuesToShow } from '../../../../common/methods/prepareInfoValuesToShow';
 import { IHederaTokenServiceKeyType } from '@/types/contract-interactions/HTS';
 import {
   HEDERA_BRANDING_COLORS,
