+++ system-contract-dapp-playground/src/components/contract-interaction/hts/token-create-custom/methods/NonFungibleTokenCreate.tsx
@@ -27,7 +27,7 @@ import { handleAPIErrors } from '../../../../common/methods/handleAPIErrors';
 import { useToastSuccessful } from '../../../../../hooks/useToastSuccessful';
 import { usePaginatedTxResults } from '../../../../../hooks/usePaginatedTxResults';
 import { SharedSigningKeysComponent } from '../../shared/components/SigningKeysForm';
-import { TransactionResultTable } from '../../shared/components/TransactionResultTable';
+import { TransactionResultTable } from '../../../../common/components/TransactionResultTable';
 import { handleSanitizeHederaFormInputs } from '../../../../common/methods/handleSanitizeFormInputs';
 import { CONTRACT_NAMES, HEDERA_TRANSACTION_RESULT_STORAGE_KEYS } from '@/utils/common/constants';
 import { useUpdateTransactionResultsToLocalStorage } from '../../../../../hooks/useUpdateLocalStorage';
